Add full support for timestamps
This commit allows date and times to be parsed and represented safely.
Because dates and times are so flexible in CE and cannot always be
represented by pre-existing types (for example, `datetime.datetime` does
not support years before 1 AD or after 999999 AD), I took it as a sign
that the project needed to introduce its own custom types to faciliate
dual encoding and decoding.

This commit therefore additionally outlines a small type system that
functions as validating wrappers around input values. These types act as
safe containers, rejecting CTE input that does not comply with CE
specifications. The idea is that we can eventually just add decoding
mechanisms to these types that take the current state and export a
CBE versions.

I've implemented `Time` and `Timestamp` as examples, and in `docs` have
supplied an outline of more type systems. For the moment, the scope is
*just* to support CTE decoding into native Python objects - later, CBE
encoding / CBE decoding / CTE encoding will be added to all.