Rewrite float parsing to use Decimal
I realized the float implementation did not support arbitrary-precision decimal
floats as mandated by the spec, so updated it to return Decimal types.
For binary floats that Decimal doesn't support, I implemented a
BinaryFloat type that just takes it as-is. Likely, this will need
updating once we get to CBE implementation.

I also ensured the Map type now accepts only valid types, and reverted
the List implementation (because it's really unnecessary)